Separates LowpassFilter and SampleSource.
