+++ Components/AY38910/AY38910.hpp
@@ -9,7 +9,7 @@
 #ifndef AY_3_8910_hpp
 #define AY_3_8910_hpp
 
-#include "../../Outputs/Speaker/Implementation/FilteringSpeaker.hpp"
+#include "../../Outputs/Speaker/Implementation/SampleSource.hpp"
 #include "../../Concurrency/AsyncTaskQueue.hpp"
 
 namespace GI {
