+++ Machines/Atari2600/TIASound.hpp
@@ -9,7 +9,7 @@
 #ifndef Atari2600_TIASound_hpp
 #define Atari2600_TIASound_hpp
 
-#include "../../Outputs/Speaker/Implementation/FilteringSpeaker.hpp"
+#include "../../Outputs/Speaker/Implementation/SampleSource.hpp"
 #include "../../Concurrency/AsyncTaskQueue.hpp"
 
 namespace Atari2600 {
