--- Outputs/Speaker/Implementation/FilteringSpeaker.cpp
@@ -1,9 +1,0 @@
-//
-//  FilteringSpeaker.cpp
-//  Clock Signal
-//
-//  Created by Thomas Harte on 15/12/2017.
-//  Copyright © 2017 Thomas Harte. All rights reserved.
-//
-
-#include "FilteringSpeaker.hpp"
