+++ Outputs/Speaker/Implementation/LowpassSpeaker.hpp
@@ -15,35 +15,16 @@
 #include "../../../ClockReceiver/ClockReceiver.hpp"
 #include "../../../Concurrency/AsyncTaskQueue.hpp"
 
+#include <cstring>
+
 namespace Outputs {
 namespace Speaker {
 
 /*!
-	A sample source is something that can provide a stream of audio.
-	This optional base class provides the interface expected to be exposed
-	by the template parameter to LowpassSpeaker.
-*/
-class SampleSource {
-	public:
-		/*!
-			Should write the next @c number_of_samples to @c target.
-		*/
-		void get_samples(std::size_t number_of_samples, int16_t *target) {}
-
-		/*!
-			Should skip the next @c number_of_samples. Subclasses of this SampleSource
-			need not implement this if it would no more efficient to do so than it is
-			merely to call get_samples and throw the result away, as per the default
-			implementation below.
-		*/
-		void skip_samples(const std::size_t number_of_samples) {
-			int16_t scratch_pad[number_of_samples];
-			get_samples(number_of_samples, scratch_pad);
-		}
-};
-
-/*!
-	The low-pass speaker
+	The low-pass speaker expects an Outputs::Speaker::SampleSource-derived
+	template class, and uses the instance supplied to its constructor as the
+	source of a high-frequency stream of audio which it filters down to a
+	lower-frequency output.
 */
 template <typename T> class LowpassSpeaker: public Speaker {
 	public:
@@ -156,9 +137,9 @@ template <typename T> class LowpassSpeaker: public Speaker {
 						uint64_t steps = stepper_->step();
 						if(steps < input_buffer_.size()) {
 							int16_t *input_buffer = input_buffer_.data();
-							memmove(	input_buffer,
-										&input_buffer[steps],
-										sizeof(int16_t) * (input_buffer_.size() - steps));
+							std::memmove(	input_buffer,
+											&input_buffer[steps],
+											sizeof(int16_t) * (input_buffer_.size() - steps));
 							input_buffer_depth_ -= steps;
 						} else {
 							if(steps > input_buffer_.size())
